+++ apps/namehashlabs.org/components/2 - molecules/footer.tsx
@@ -46,7 +46,7 @@ const footerResources = [
 export const Footer = (props: React.HTMLAttributes<HTMLDivElement>) => {
   return (
     <section
-      className="lg:px-[112px] px-5 flex items-center justify-center w-full border-t border-gray-200"
+      className="lg:px-[50px] px-5 flex items-center justify-center w-full border-t border-gray-200"
       {...props}
     >
       <div className="pt-8 pb-5 flex flex-col gap-10 items-start justify-between w-full max-w-[1216px]">