+++ apps/namekit.io/src/components/atoms/icons/explore-web3-lp/pattern.tsx
@@ -1,0 +1,34 @@
+import { type SVGProps } from "react";
+
+export const DotPattern = (props: SVGProps<SVGSVGElement>) => {
+  return (
+    <svg width="100%" height="100%" {...props}>
+      <pattern
+        id="pattern-circles"
+        x="0"
+        y="0"
+        width="20"
+        height="20"
+        patternUnits="userSpaceOnUse"
+        patternContentUnits="userSpaceOnUse"
+      >
+        <circle
+          id="pattern-circle"
+          cx="10"
+          cy="10"
+          r="1.6257413380501518"
+          fill="#000"
+        ></circle>
+      </pattern>
+
+      <rect
+        id="rect"
+        x="0"
+        y="0"
+        width="100%"
+        height="100%"
+        fill="url(#pattern-circles)"
+      ></rect>
+    </svg>
+  );
+};