+++ apps/namekit.io/src/components/atoms/icons/info-bordered-icon.tsx
@@ -1,0 +1,22 @@
+import { type SVGProps } from "react";
+
+export const InfoBorderedIcon = (props: SVGProps<SVGSVGElement>) => {
+  return (
+    <svg
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M11 14H10V10H9M10 6H10.01M19 10C19 14.9706 14.9706 19 10 19C5.02944 19 1 14.9706 1 10C1 5.02944 5.02944 1 10 1C14.9706 1 19 5.02944 19 10Z"
+        stroke="#808080"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+
+export default InfoBorderedIcon;