+++ packages/ens-utils/src/hashutils.test.ts
@@ -2,9 +2,22 @@ import { describe, it, expect } from "vitest";
 import {
   isEncodedLabelhash,
   isKeccak256Hash,
+  labelhash,
   normalizeEncodedLabelhash,
   normalizeKeccak256Hash,
 } from "./hashutils";
+
+describe("labelhash", () => {
+  it("correctly hashes an empty label", () => {
+    const result = labelhash("");
+    expect(result).toBe("0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470");
+  });
+
+  it("correctly hashes a non-empty label", () => {
+    const result = labelhash("example");
+    expect(result).toBe("0x6fd43e7cffc31bb581d7421c8698e29aa2bd8e7186a394b85299908b4eb9b175");
+  });
+});
 
 describe("isKeccak256Hash", () => {
   it("valid Keccak256Hash: with prefix all lowercase", () => {