+++ internal/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @namekit/tsconfig
 
+## 0.2.0
+
+### Minor Changes
+
+- 1891ce7: release
+
 ## 0.1.0
 
 ### Minor Changes