+++ packages/nameguard-react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @namehash/nameguard-react
 
+## 0.8.0
+
+### Minor Changes
+
+- 1891ce7: release
+
+### Patch Changes
+
+- Updated dependencies [1891ce7]
+- Updated dependencies [c748998]
+  - @namehash/ens-utils@1.19.0
+  - @namehash/ens-webfont@0.7.0
+  - @namehash/nameguard@0.8.0
+  - @namehash/namekit-react@0.10.0
+
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [47de3ad]
+- Updated dependencies [d5183f1]
+  - @namehash/nameguard@0.7.0
+  - @namehash/namekit-react@0.9.0
+
 ## 0.7.1
 
 ### Patch Changes